LPS-199140 Transform class SharedAssetsFilterItemRegistry to class util, move class SharingEntryDropdownItemContributorRegistry to class ViewSharedAssetsDisplayContext, and apply to usages
