+++ modules/apps/sharing/sharing-web/src/main/java/com/liferay/sharing/web/internal/display/context/ViewSharedAssetsManagementToolbarDisplayContext.java
@@ -22,8 +22,8 @@ import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.sharing.filter.SharedAssetsFilterItem;
-import com.liferay.sharing.web.internal.filter.SharedAssetsFilterItemRegistry;
 import com.liferay.sharing.web.internal.item.selector.SharedAssetsFilterItemItemSelectorCriterion;
+import com.liferay.sharing.web.internal.util.SharedAssetsFilterItemRegistryUtil;
 
 import java.util.List;
 import java.util.Map;
@@ -45,7 +45,6 @@ public class ViewSharedAssetsManagementToolbarDisplayContext
 		LiferayPortletRequest liferayPortletRequest,
 		LiferayPortletResponse liferayPortletResponse,
 		SearchContainer<?> searchContainer,
-		SharedAssetsFilterItemRegistry sharedAssetsFilterItemRegistry,
 		ViewSharedAssetsDisplayContext viewSharedAssetsDisplayContext) {
 
 		super(
@@ -53,7 +52,6 @@ public class ViewSharedAssetsManagementToolbarDisplayContext
 			searchContainer);
 
 		_itemSelector = itemSelector;
-		_sharedAssetsFilterItemRegistry = sharedAssetsFilterItemRegistry;
 		_viewSharedAssetsDisplayContext = viewSharedAssetsDisplayContext;
 	}
 
@@ -144,7 +142,7 @@ public class ViewSharedAssetsManagementToolbarDisplayContext
 	private String _getClassNameLabel(String className) {
 		if (Validator.isNotNull(className)) {
 			SharedAssetsFilterItem sharedAssetsFilterItem =
-				_sharedAssetsFilterItemRegistry.getSharedAssetsFilterItem(
+				SharedAssetsFilterItemRegistryUtil.getSharedAssetsFilterItem(
 					className);
 
 			if (sharedAssetsFilterItem != null) {
@@ -201,8 +199,6 @@ public class ViewSharedAssetsManagementToolbarDisplayContext
 	}
 
 	private final ItemSelector _itemSelector;
-	private final SharedAssetsFilterItemRegistry
-		_sharedAssetsFilterItemRegistry;
 	private final ViewSharedAssetsDisplayContext
 		_viewSharedAssetsDisplayContext;
