+++ modules/apps/sharing/sharing-web/src/main/java/com/liferay/sharing/web/internal/item/selector/SharedAssetsFilterItemItemSelectorView.java
@@ -13,7 +13,7 @@ import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.sharing.filter.SharedAssetsFilterItem;
-import com.liferay.sharing.web.internal.filter.SharedAssetsFilterItemRegistry;
+import com.liferay.sharing.web.internal.util.SharedAssetsFilterItemRegistryUtil;
 
 import java.io.IOException;
 
@@ -127,7 +127,7 @@ public class SharedAssetsFilterItemItemSelectorView
 							portletURL, null, null);
 
 					List<SharedAssetsFilterItem> sharedAssetsFilterItems =
-						_sharedAssetsFilterItemRegistry.
+						SharedAssetsFilterItemRegistryUtil.
 							getSharedAssetsFilterItems();
 
 					entriesSearchContainer.setResultsAndTotal(
@@ -161,8 +161,5 @@ public class SharedAssetsFilterItemItemSelectorView
 
 	@Reference
 	private Language _language;
-
-	@Reference
-	private SharedAssetsFilterItemRegistry _sharedAssetsFilterItemRegistry;
 
 }
